// frontend/lib/generated-api/models/ReferenceFetchConclusion.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Analyst API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const ReferenceFetchConclusion = {
  SourceFound: 'source_found',
  SourceNotFound: 'source_not_found',
  SourceFoundButNotAccessible: 'source_found_but_not_accessible',
} as const;
export type ReferenceFetchConclusion = (typeof ReferenceFetchConclusion)[keyof typeof ReferenceFetchConclusion];

export function instanceOfReferenceFetchConclusion(value: any): boolean {
  for (const key in ReferenceFetchConclusion) {
    if (Object.prototype.hasOwnProperty.call(ReferenceFetchConclusion, key)) {
      if (ReferenceFetchConclusion[key as keyof typeof ReferenceFetchConclusion] === value) {
        return true;
      }
    }
  }
  return false;
}

export function ReferenceFetchConclusionFromJSON(json: any): ReferenceFetchConclusion {
  return ReferenceFetchConclusionFromJSONTyped(json, false);
}

export function ReferenceFetchConclusionFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): ReferenceFetchConclusion {
  return json as ReferenceFetchConclusion;
}

export function ReferenceFetchConclusionToJSON(value?: ReferenceFetchConclusion | null): any {
  return value as any;
}

export function ReferenceFetchConclusionToJSONTyped(
  value: any,
  ignoreDiscriminator: boolean,
): ReferenceFetchConclusion {
  return value as ReferenceFetchConclusion;
}